Move router folder to lib folder
