+++ lib/di/components/service_locator.config.dart
@@ -12,9 +12,9 @@ import 'package:get_it/get_it.dart' as _i1;
 import 'package:get_storage/get_storage.dart' as _i4;
 import 'package:injectable/injectable.dart' as _i2;
 
-import '../../config/router/app_router.dart' as _i6;
 import '../../data/getstore/get_store_helper.dart' as _i3;
 import '../../data/hive/hive_helper.dart' as _i5;
+import '../../router/app_router.dart' as _i6;
 
 extension GetItInjectableX on _i1.GetIt {
 // initializes the registration of main-scope dependencies inside of GetIt
