+++ lib/router/app_router.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:injectable/injectable.dart';
 
-import '../../data/getstore/get_store_helper.dart';
-import '../../di/components/service_locator.dart';
-import '../../features/home/home.dart';
-import '../../features/info/info_screen.dart';
+import '../data/getstore/get_store_helper.dart';
+import '../di/components/service_locator.dart';
+import '../features/home/home.dart';
+import '../features/info/info_screen.dart';
 import 'fade_extension.dart';
 
 GetStoreHelper getStoreHelper = getIt<GetStoreHelper>();
